% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity.test.R
\name{homogeneity.test}
\alias{homogeneity.test}
\title{Test of group homogeneity.}
\usage{
homogeneity.test(md, numB)
}
\arguments{
\item{md}{A n x n matrix of distances where \eqn{n = n1 + n2}.}

\item{numB}{Numeric scalar. The number of bootstraps. It's recommended \eqn{1000 < numB < 10000}.}
}
\value{
The smallest Utest's p-value of all (\eqn{(2^(n - 1) - n - 1)}) two-subgroups arrangement.
}
\description{
If the smallest p-value is greater than \code{alpha}, then the group is considered homogeneous.
}
\examples{
# Here n = 11 and the group is not homogeneous.
x <- matrix(c(rnorm(25, 0), rnorm(30, 10)), nrow = 5)
md <- as.matrix(dist(t(x)))
homogeneity.test(md, 1000)

# Here n = 11 and the group is homogeneous.
x <- matrix(c(rnorm(25, 0), rnorm(30, 0)), nrow = 5)
md <- as.matrix(dist(t(x)))
homogeneity.test(md, 1000)
}
