% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utest_classify.R
\name{print.utest_classify}
\alias{print.utest_classify}
\title{Simple print method for utest_classify objects.}
\usage{
\method{print}{utest_classify}(obj, ...)
}
\description{
Simple print method for utest_classify objects.
}
