% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BnClustSig.R
\name{rep_optimBn}
\alias{rep_optimBn}
\title{Optimization function with multiple staring points (for local optima)}
\usage{
rep_optimBn(mdm, rep = 15, bootB = -1)
}
\arguments{
\item{mdm}{Matrix of squared Euclidean distances between all data points.}

\item{rep}{Number of replications}

\item{bootB}{Result of previous bootstrap (if available). If, -1, a new bootstrap is performed for the variance of Bn.}
}
\description{
Finds the configuration with max Bn among all configurations.
}
