% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utest.R
\name{Utest}
\alias{Utest}
\title{Test for homogeneity of two groups.}
\usage{
Utest(ngv, numB, md)
}
\arguments{
\item{ngv}{A numeric vector \code{c(n1, n2)} where \code{n1} is the number of elements in group 1 and
\code{n2} is the number of elements in group 2.}

\item{numB}{Numeric scalar. The number of bootstraps. It's recommended \eqn{1000 < numB < 10000}.}

\item{md}{A n x n matrix of distances where \eqn{n = n1 + n2}.}
}
\value{
p-value based on bootstraps.
}
\description{
Test for homogeneity of two groups.
}
