% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utest_class.R
\name{Uteste_class}
\alias{Uteste_class}
\title{Test for classification in one of two groups.}
\usage{
Uteste_class(ngv, numB, md)
}
\arguments{
\item{ngv}{A numeric vector \code{c(n1, n2)} where \code{n1} is the number of elements in group 1 and
\code{n2} is the number of elements in group 2.}

\item{numB}{Numeric scalar. The number of bootstraps. It's recommended \eqn{1000 < numB < 10000}.}

\item{md}{A \eqn{(n+1) x (n+1)} matrix of distances where \eqn{n = n1 + n2}.}
}
\value{
p-value based on bootstraps.
}
\description{
The null hypotheses is that the new data is not well classified and the alternative hypotheses is that
the data is well classified.
}
\details{
The data in \code{md} must be in order: the first \code{n1} observations must belong to group 1,
then the new observation to be classified and the next \code{n2} observations must belong to group 2.
}
\examples{
# Example 1
# x is a 60 x 11 matrix.
x <- matrix(c(rnorm(300, 0), rnorm(360, 10)), nrow = 60)
# Five columns with mean zero and six with mean 10.
md <- as.matrix(dist(t(x)))
ngv <- c(5, 5)
# H0: column 6 is not significantly classified in group 1 (columns 1, 2, 3, 4, 5) (TRUE p-value must be >0.05)
Uteste_class(ngv, 1000, md)

# Example 2
# x is a 60 x 11 matrix.
x <- matrix(c(rnorm(360, 0), rnorm(300, 10)), nrow = 60)
# Six columns with mean zero and five with mean 10.
md <- as.matrix(dist(t(x)))
ngv <- c(5, 5)
# H0: column 6 is not significantly classified in group 1 (columns 1, 2, 3, 4, 5) (FALSE p-value must be >0.05)
Uteste_class(ngv, 1000, md)
}
