% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_sigma.R
\name{boot_sigma}
\alias{boot_sigma}
\title{Bootstrap to find the variance of \code{Bn}.}
\usage{
boot_sigma(ngv, numB, md)
}
\value{
The variance of \code{Bn}.
}
\description{
Bootstrap to find the variance of \code{Bn}.
}
