% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimBn.R
\name{optimBn}
\alias{optimBn}
\title{Optimization of standardized \code{Bn}.}
\usage{
optimBn(dados, itmax = 200, centers = -1)
}
\arguments{
\item{dados}{Numeric vector.}

\item{itmax}{Numeric scalar. The maximum number of iterations.}

\item{centers}{Numeric scalar. The number of centers.}
}
\value{
The group to which the objective function based on standardized \code{Bn}
had been converged to a minimum.
}
\description{
Optimization of standardized \code{Bn}.
}
