% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utest_class.R
\name{print.utest_classify}
\alias{print.utest_classify}
\title{Simple print method for utest_classify objects.}
\usage{
\method{print}{utest_classify}(x, ...)
}
\arguments{
\item{x}{utest_classify object}

\item{...}{additional parameters passed to the function}
}
\description{
Simple print method for utest_classify objects.
}
